#267. Fix ios crash when making backend request.
