--- common/src/commonMain/kotlin/io/xorum/codeforceswatcher/network/responses/backend/Response.kt
@@ -1,12 +1,0 @@
-package io.xorum.codeforceswatcher.network.responses.backend
-
-sealed class Response<T> {
-
-    data class Success<T>(
-            val result: T
-    ) : Response<T>()
-
-    data class Failure<T>(
-            val error: String?
-    ) : Response<T>()
-}
